+++ malachite-nz/src/bin_util/demo_and_bench/natural_polynomial/arithmetic/mod.rs
@@ -11,7 +11,9 @@ use malachite_base::test_util::runner::Runner;
 pub(crate) fn register(runner: &mut Runner) {
     height::register(runner);
     mod_is_reduced::register(runner);
+    mod_power_of_2::register(runner);
 }
 
 mod height;
 mod mod_is_reduced;
+mod mod_power_of_2;